pyproject.toml
[project]
name = "onclaw"
version = "0.1.0"
description = "Kubernetes alert investigation bot for Slack and Telegram"
requires-python = ">=3.11"
dependencies = [
    "slack-bolt>=1.27,<2",
    "slack-sdk>=3.33,<4",
    "kubernetes>=31,<36",
    "anthropic>=0.84,<1",
    "pyyaml>=6.0,<7",
    "pydantic>=2.0,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov",
    "ruff",
    "mypy",
]

[project.scripts]
onclaw = "onclaw.__main__:main"

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
